requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.3.0

# OCR dependencies
easyocr>=1.6.0
pytesseract>=0.3.10
opencv-python>=4.5.0

# Translation dependencies
transformers>=4.20.0
torch>=1.12.0
sentencepiece>=0.1.96

# Q&A system dependencies
faiss-cpu>=1.7.0
langchain>=0.0.200
sentence-transformers>=2.2.0
pypdf2>=2.0.0
python-docx>=0.8.11

# Forecasting dependencies
prophet>=1.1.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Backend dependencies (optional)
fastapi>=0.78.0
uvicorn>=0.17.0
python-multipart>=0.0.5

# Frontend dependencies (optional)
streamlit>=1.22.0

# MLflow (optional)
mlflow>=1.28.0

# Utility dependencies
pyyaml>=6.0
requests>=2.28.0
tqdm>=4.64.0